requirements.txt
streamlit==1.32.0
youtube-transcript-api==0.6.2
requests==2.31.0
python-dotenv==1.0.0
openai==0.28.1
tiktoken==0.5.2